+++ src/App.tsx
@@ -1,12 +1,12 @@
 import { lazy, Suspense } from "react";
 
 import GlobalStyle from "@/components/GlobalStyle";
-import { Section1, Section2 } from "@/components/sections/";
+import { Section1, Section2 } from "@/features";
 
-const Section3 = lazy(() => import("@/components/sections/section3/Section3"));
-const Section4 = lazy(() => import("@/components/sections/section4/Section4"));
-const Footer = lazy(() => import("@/components/Footer"));
-const Menu = lazy(() => import("@/components/sections/menu/Menu"));
+const Section3 = lazy(() => import("@/features/Section3"));
+const Section4 = lazy(() => import("@/features/Section4"));
+const Footer = lazy(() => import("@/features/Footer"));
+const Menu = lazy(() => import("@/features/Menu"));
 
 function App() {
   return (
