--- src/components/sections/index.ts
@@ -1,5 +1,0 @@
-export { default as Menu } from "./menu/Menu";
-export { default as Section1 } from "./section1/Section1";
-export { default as Section2 } from "./section2/Section2";
-export { default as Section3 } from "./section3/Section3";
-export { default as Section4 } from "./section4/Section4";
