+++ src/features/Section1/ScrollIndicator.tsx
