+++ src/features/Section1/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Section1";
