+++ src/features/Section2/TechChip.tsx
