+++ src/features/Section2/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Section2";
