+++ src/features/Section4/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Section4";
