+++ src/features/index.ts
@@ -1,0 +1,5 @@
+export { default as Menu } from "./Menu";
+export { default as Section1 } from "./Section1/Section1";
+export { default as Section2 } from "./Section2/Section2";
+export { default as Section3 } from "./Section3/Section3";
+export { default as Section4 } from "./Section4/Section4";
